move sdsf only tests to examples
